Create PIN pad on the scene and make the PIN pad can work with the Gaze system.
Canvas object is the background of the PIN pad. And there are nine textbox on the canvas. Each textbox is one of the button on the PIN pad.
As discribed on the API document (https://vr.tobii.com/sdk/develop/unity/documentation/usage-examples/), there is a script attached under each textbox to make textboxes gazefocusable and can interact with gaze system.